{"types":"C","runs":15}

+++ CountSequence.java
@@ -22,7 +22,7 @@ public class CountSequence extends ConsoleProgram{
 	private RandomGenerator rgen = RandomGenerator.getInstance();
 	
 	public void run() {
-		n =100;
+		n = 100;
 		int sumExp = 1;
 		for(int i = 0; i < NUM_EXP; i++){
 			while(n!=0){
@@ -33,7 +33,7 @@ public class CountSequence extends ConsoleProgram{
 				}
 			}
 		}
-		int avgSum = sumExp / NUM_EXP;
+		double avgSum = sumExp / NUM_EXP;
 		println(avgSum);
 	}
 }
